- Création du cogs file.py - Mise en commun des fonctions importantes dans Utils/utils.py - Modification dans app.py pour que ça marche avec le system
